Converting to new prop g/s/d
The new method allows getters setters and deleters to easily be subclassed out. This is one of the really important things I want with this library. Is for functionality to be easily subclassed. The reason for this is it incourages people to create their own custom classes to be reused and that's the point of WPL to be DRY.